refactor(graphql): create graphql directives test trait
