+++ packages/composer/amazeelabs/graphql_directives/tests/assets/utilities/.graphqlconfig
@@ -1,0 +1,5 @@
+{
+  "name": "Directives Test - Utilities",
+  "schemaPath": "./schema.graphqls",
+  "includes": ["./directives.graphqls", "./extension/*.graphqls"]
+}
