+++ packages/composer/amazeelabs/graphql_directives/tests/assets/utilities/extension
@@ -1,0 +1,1 @@
+../../modules/graphql_directives_test/graphql
